MAINT: restructure to avoid import cycles
